requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
simple-websocket==1.0.0
PyYAML==6.0.1